requirements.txt
newspaper3k
nltk
lxml_html_clean
pymysql
DBUtils
pytz
slack_sdk
python-dotenv
googletrans==4.0.0rc1